requirements.txt
arabic-reshaper==3.0.0
asgiref==3.9.1
asn1crypto==1.5.1
boto3==1.38.46
botocore==1.38.46
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
cssselect2==0.8.0
diff-match-patch==20241021
Django==5.2.5
django-appconf==1.1.0
django-ckeditor==6.7.3
django-cleanup==9.0.0
django-cors-headers==4.7.0
django-crontab==0.7.1
django-debug-toolbar==5.2.0
django-easy-audit==1.3.7
django-environ==0.12.0
django-filter==25.1
django-import-export==4.3.8
django-js-asset==3.1.2
django-mathfilters==1.0.0
django-multiselectfield==1.0.1
django-rest-framework==0.1.0
django-select2==8.4.1
django-storages==1.14.6
djangorestframework==3.16.0
docopt==0.6.2
et_xmlfile==2.0.0
gunicorn==23.0.0
html5lib==1.1
idna==3.10
jmespath==1.0.1
lxml==6.0.0
mysqlclient==2.2.7
num2words==0.5.14
numpy==2.3.1
openpyxl==3.1.5
oscrypto==1.3.0
packaging==25.0
pandas==2.3.0
pillow==11.2.1
pycparser==2.22
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
pypdf==5.6.1
pypng==0.20220715.0
PyQRCode==1.2.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
reportlab==4.4.2
requests==2.32.4
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tablib==3.8.0
tinycss2==1.4.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
webencodings==0.5.1
xhtml2pdf==0.2.17
xlsxwriter==3.2.5
xmltodict==0.14.2
